Further refined rogue bashing
I added a balestra toggle to the system. Turns out that with enough
critical rating at a high level. Just using balestra instead of thrust
is enough to bash effectively. This holds true when wielding a
longsword, which balestra's speed is not effected much by it (very
similar to how obliterate works for soldiers).