Hiding toggle feature update
The guile checks for hiding on prompt updates is creates a lot of spam
and can be inconvenient in out of combat situations, such as in a
timewarp prepped room. Adjusted the hiding feature of the system to hide
at 80% guile only when person is checking info here after bashing.