Complete Grip Changes for Priest and Rogues
This should be the final changes to bashing checks for rogues and
priests. It will check for whether or not they have the grip defense
instead of checking if they have the skill since you can't get the grip
skill if you are a non-soldier class. If you are not toggling the
gripping (please make sure the default for matawa's gripping toggle is
false), then the system should still be able to move forward with the
bashing function instead of stopping because they didn't have the
gripping skill within a non-soldier class like before (Many novices
complained to me that their rogues/priests can't bash cause the function
has a skill check for grip).